pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acme-multi-agent"
version = "0.1.0"
description = "Multi-Agent Routing System for Acme Corporation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Acme Engineering"}
]
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.7.0",
    "openai>=1.0.0",
    "langfuse-langchain>=2.0.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.0.0",
    "jupyter>=1.0.0",
    "pytest>=7.0.0",
]

[project.scripts]
acme-assistant = "src.multi_agent_system:main"
acme-index = "src.indexing:main"
acme-evaluate = "src.evaluator:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
